flask_heroku2/requirements.txt
certifi==2022.5.18.1
click==8.1.3
colorama==0.4.5
Flask==2.1.2
gunicorn==20.1.0
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.1.2
wincertstore==0.2
zipp==3.8.0
category-encoders==2.5.0
lightgbm==3.3.2
numpy==1.23.0
pandas==1.4.2
patsy==0.5.2
psycopg2-binary==2.9.3
scikit-learn==1.1.1
scipy==1.8.1
pickle-mixin==1.0.2
